+++ skyux-spa-visual-tests/src/app/home.component.ts
@@ -1,0 +1,44 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'my-home',
+  templateUrl: './home.component.html'
+})
+export class HomeComponent {
+
+  public tests: Array<string> = [
+    'action-button',
+    'alert',
+    'avatar',
+    'button',
+    'card',
+    'checkbox',
+    'chevron',
+    'datepicker',
+    'definition-list',
+    'dropdown',
+    'error',
+    'fileattachments',
+    'filter',
+    'grid',
+    'key-info',
+    'label',
+    'list-secondary-actions',
+    'list-view-checklist',
+    'modal',
+    'navbar',
+    'page-summary',
+    'paging',
+    'radio',
+    'repeater',
+    'search',
+    'sort',
+    'tabs',
+    'text-expand',
+    'text-expand-repeater',
+    'text-highlight',
+    'tiles',
+    'toolbar',
+    'wait'
+  ];
+}