+++ src/modules/dropdown/dropdown.component.spec.ts
@@ -104,6 +104,32 @@ describe('Dropdown component', () => {
 
       expect(dropdownMenu1).not.toBeVisible();
     });
+
+     it('should display default label when label not set', () => {
+
+      let fixture = TestBed.createComponent(DropdownParentTestComponent);
+      let el: HTMLElement = fixture.nativeElement;
+
+      fixture.detectChanges();
+
+      let button = el.querySelector('#dropdown-1 .sky-dropdown-button') as HTMLButtonElement;
+      let label = button.getAttribute('aria-label');
+
+      expect(label).toBe('Context menu');
+    });
+
+     it('should display default label when label is set', () => {
+
+      let fixture = TestBed.createComponent(DropdownParentTestComponent);
+      let el: HTMLElement = fixture.nativeElement;
+
+      fixture.detectChanges();
+
+      let button = el.querySelector('#dropdown-4 .sky-dropdown-button') as HTMLButtonElement;
+      let label = button.getAttribute('aria-label');
+
+      expect(label).toBe('test label');
+    });
   });
 
   describe('postition tests', () => {
@@ -333,7 +359,7 @@ describe('Dropdown component', () => {
       expect(getDropdownBtnEl(el)).toHaveCssClass('sky-dropdown-button-type-context-menu');
     });
 
-   it('should have a default button background of "sky-btn-default" aka isPrimary is false', () => {
+   it('should have a default button background of "sky-btn-default"', () => {
       let fixture = TestBed.createComponent(DropdownTestComponent);
       let el: Element = fixture.nativeElement;
 
@@ -342,12 +368,12 @@ describe('Dropdown component', () => {
       expect(getDropdownBtnEl(el)).toHaveCssClass('sky-btn-default');
     });
 
-    it('should set the CSS class to primary when isPrimary is true', () => {
+    it('should set the CSS class based on buttonStyle changes', () => {
       let fixture = TestBed.createComponent(DropdownTestComponent);
       let cmp = fixture.componentInstance;
       let el: Element = fixture.nativeElement;
 
-      cmp.isPrimary = true;
+      cmp.buttonStyle = 'primary';
 
       fixture.detectChanges();