+++ src/modules/dropdown/dropdown.component.ts
@@ -9,6 +9,7 @@ import {
 import { SkyDropdownAdapterService } from './dropdown-adapter.service';
 
 import { SkyWindowRefService } from '../window';
+import { SkyResources } from '../resources';
 
 @Component({
   selector: 'sky-dropdown',
@@ -38,7 +39,13 @@ export class SkyDropdownComponent implements OnDestroy {
   }
 
   @Input()
-  public label: string;
+  public get label(): string{
+    return this._label || SkyResources.getString('context_menu_default_label');
+  }
+
+  public set label(value: string) {
+    this._label = value;
+  }
 
   @Input()
   public title: string;
@@ -47,12 +54,12 @@ export class SkyDropdownComponent implements OnDestroy {
   public alignment: string = 'left';
 
   @Input()
-  public get isPrimary(): boolean {
-    return this._isPrimary;
+  public get buttonStyle(): string{
+    return this._buttonStyle || 'default';
   }
 
-  public set isPrimary(flag: boolean) {
-    this._isPrimary = flag;
+  public set buttonStyle(value: string) {
+    this._buttonStyle = value;
   }
 
   private open = false;
@@ -61,9 +68,11 @@ export class SkyDropdownComponent implements OnDestroy {
 
   private _buttonType: string;
 
-  private _isPrimary: boolean;
+  private _buttonStyle: string;
 
   private _trigger: string;
+
+  private _label: string;
 
   constructor(
     private renderer: Renderer,