+++ src/modules/page-summary/page-summary-image/page-summary-image.component.ts
@@ -5,4 +5,6 @@ import { Component } from '@angular/core';
   templateUrl: './page-summary-image.component.html',
   styleUrls: ['./page-summary-image.component.scss']
 })
+/* istanbul ignore next */
+/* Code coverage having problems with no statements in classes */
 export class SkyPageSummaryImageComponent { }